pyproject.toml
[project]
name = "intelligent-research-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "celery[redis]>=5.6.2",
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "pytest>=9.0.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
]
